Finishes day 03, part 2
I rearranged part 1 because I thought part 2's input might be so large
that I couldn't parse it in memory and would have to stream it. It looks
me a while to see that the input file was actually multiple lines and my
part 1 solution was handling that well. I also realized it was good to
test the final answer for part 1 with a check-equal? so I could be sure
I kept getting the right answer. When I removed the map, map, map layer
from part 1 I was effectively ignoring all but the first line of text so
it didn't work any more. It look me a while to realize that I was
getting part 2's answer wrong because I prematurely optimized part 1
without checking for its correctness. I just checked that it compiled
and worked on the simple (1 line) input. Instead of reverting to maps I
joined all the strings, which fixed both parts and reduced the
code (though it wouldn't work with very large inputs).

Still suffers from the run-code duplication problem.